# protocol/openapi.yaml
...
openapi: "3.0.3"
info:
  title: lightcode-protocol
  version: "1"
servers:
  - url: "http://127.0.0.1"
# Every path requires the bearer credential; it is declared once here and
# enforced by the Runtime when the server lands.
security:
  - BearerAuth: []
paths:
...
  /v1/sessions:
    get:
      operationId: ListSessions
      parameters:
        - $ref: "#/components/parameters/WorkspaceQuery"
        - $ref: "#/components/parameters/LifecycleQuery"
      responses:
        "200":
          description: Projected Session headers, one revisioned row per Session.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: CreateSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: The created Session header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}:
    delete:
      operationId: DeleteSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: The archived Session was deleted.
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/archive:
    post:
      operationId: ArchiveSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: The archived Session header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/reopen:
    post:
      operationId: ReopenSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: The reopened Session header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/submit:
    post:
      operationId: SubmitSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitRequest"
      responses:
        "200":
          description: The submission disposition and Operation when admitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitResult"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/compact:
    post:
      operationId: CompactSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompactRequest"
      responses:
        "200":
          description: The admitted compaction Operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/interrupt:
    post:
      operationId: InterruptSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: The interrupt control was issued; it does not wait for the terminal commit.
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/stop:
    post:
      operationId: StopSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: The stop control completed its background convergence wait.
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/fork:
    post:
      operationId: ForkSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForkRequest"
      responses:
        "200":
          description: The forked Session and its admitted Operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForkResult"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/agent-type:
    put:
      operationId: SetSessionAgentType
      parameters:
        - $ref: "#/components/parameters/SessionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetSessionAgentTypeRequest"
      responses:
        "200":
          description: The Session header; only subsequent admissions use the new Agent type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/hydration:
    get:
      operationId: GetSessionHydration
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: One coherent Session projection captured at one revision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hydration"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/history:
    get:
      operationId: GetSessionHistory
      parameters:
        - $ref: "#/components/parameters/SessionID"
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: One anchored page of conversation items in ascending display order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/code-snapshots:
    get:
      operationId: GetSessionCodeSnapshots
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: Per-Operation code snapshot groups, ascending by admitted entry.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodeSnapshotGroup"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{id}/revert-code:
    post:
      operationId: RevertSessionCode
      parameters:
        - $ref: "#/components/parameters/SessionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevertCodeRequest"
      responses:
        "200":
          description: The completed restore result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeRevertResult"
        default:
          description: >
            A traversal I/O failure after earlier groups changed returns the
            same result DTO with accumulated restored/skipped members and the
            error member; every other failure, including a pre-traversal
            storage failure, returns the schema error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeRevertError"
  /v1/workspaces:
    get:
      operationId: ListWorkspaces
      responses:
        "200":
          description: Distinct normalized Workspace roots derived from validated Sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
        default:
          $ref: "#/components/responses/Error"
  /v1/workspaces/files/read:
    post:
      operationId: ReadWorkspaceFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadFileRequest"
      responses:
        "200":
          description: The bounded file view.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadFileResult"
        default:
          $ref: "#/components/responses/Error"
  /v1/retained-code-snapshots:
    get:
      operationId: GetRetainedCodeSnapshots
      parameters:
        - name: workspace
          in: query
          required: true
          schema:
            type: string
        - name: session_id
          in: query
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{8}$"
      responses:
        "200":
          description: The legacy per-turn snapshot groups of one pre-cutover Session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetainedTurns"
        default:
          $ref: "#/components/responses/Error"
  /v1/retained-code-snapshots/revert:
    post:
      operationId: RevertRetainedCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetainedRevertRequest"
      responses:
        "200":
          description: The completed legacy restore result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeRevertResult"
        default:
          description: >
            A traversal I/O failure after earlier turns changed returns the
            same result DTO with accumulated restored/skipped members and the
            error member; every other failure, including a pre-traversal
            storage failure, returns the schema error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeRevertError"
  /v1/configuration:
    get:
      operationId: GetConfiguration
      responses:
        "200":
          description: The revisioned configuration view.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigurationView"
        default:
          $ref: "#/components/responses/Error"
  /v1/configuration/reload:
    post:
      operationId: ReloadConfiguration
      responses:
        "200":
          description: A publication, not an edit — no result member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/configuration/settings:
    put:
      operationId: UpdateConfigurationSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: The published settings under the mutation envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsMutation"
        default:
          $ref: "#/components/responses/Error"
  /v1/agent-types/model:
    put:
      operationId: SetAgentTypeModel
      parameters:
        - name: agent_type
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAgentTypeModelRequest"
      responses:
        "200":
          description: The updated Agent roster entry under the mutation envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMutation"
        default:
          $ref: "#/components/responses/Error"
  /v1/models:
    get:
      operationId: ListModels
      parameters:
        - name: all
          in: query
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: >
            The full or visible flat model list, preserving the retained
            connected-only and hidden rules.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelList"
        default:
          $ref: "#/components/responses/Error"
  /v1/providers:
    get:
      operationId: ListProviders
      responses:
        "200":
          description: Every provider view, listed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderList"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: CreateProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderRequest"
      responses:
        "200":
          description: The added provider view; a generated key name is its generated_key_env.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderMutation"
        default:
          $ref: "#/components/responses/Error"
...
  /v1/providers/discover:
    post:
      operationId: DiscoverProviderCandidates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoveryRequest"
      responses:
        "200":
          description: Discovered model candidates, persisted nowhere.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscoveredModelCandidate"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        The Runtime owner's random 32-byte credential encoded as 64 lowercase
        hex, sent as "Authorization: Bearer <credential>" on every path.
  parameters:
...
    WorkspaceQuery:
      name: workspace
      in: query
      required: true
      schema:
        type: string
    LifecycleQuery:
      name: lifecycle
      in: query
      required: true
      schema:
        $ref: "#/components/schemas/SessionLifecycle"
...
  responses:
    Error:
      description: One schema-defined typed error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Session:
      type: object
      additionalProperties: false
      required:
        - session_id
        - workspace
        - lifecycle
        - agent_type
        - created_at
        - last_activity
        - session_revision
      properties:
        session_id:
          type: string
          pattern: "^[0-9a-f]{32}$"
        workspace:
          type: string
        parent_session_id:
          type: string
          pattern: "^[0-9a-f]{32}$"
        source_session_id:
          type: string
          pattern: "^[0-9a-f]{32}$"
        lifecycle:
          $ref: "#/components/schemas/SessionLifecycle"
        agent_type:
          type: string
        current_operation_id:
          type: string
        created_at:
          type: string
          format: date-time
        last_activity:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time
        session_revision:
          $ref: "#/components/schemas/SessionRevision"
...